Keep IntND at user boundaries; add RealND for ndimage/argmax primitives
Revert the int64-accepting widening of `_ParamsLeaf` and the
`UserInitialConditions` alias: pylcm pins integers to int32 everywhere,
so the `process_params` / `Model.simulate` boundaries take `IntND` and
reject raw int64 fail-fast. Add `RealND` (jaxtyping `Real`, any-width
float or int) for the genuinely dtype-polymorphic `ndimage` / `argmax`
primitives, which are exercised across dtypes below the canonical-dtype
invariant. Pin internal `jnp.arange` index arrays to int32 so pylcm
never produces int64 under x64. Fix tests that passed raw int64 initial
conditions / params to use the correct dtype.

+++ src/lcm/regime_building/processing.py
@@ -868,7 +868,7 @@ def _get_stochastic_next_function_for_shock(
 
     @with_signature(args={f"{name}": "ContinuousState"}, return_annotation="Int1D")
     def next_func(**kwargs: Any) -> Int1D:  # noqa: ARG001, ANN401
-        return jnp.arange(grid.shape[0])
+        return jnp.arange(grid.shape[0], dtype=jnp.int32)
 
     return next_func
 
@@ -902,7 +902,7 @@ def _get_weights_func_for_shock(*, name: str, grid: _ShockGrid) -> UserFunction:
                 input=transition_probs,
                 coordinates=[
                     jnp.full(n_points, fill_value=coord),
-                    jnp.arange(n_points),
+                    jnp.arange(n_points, dtype=jnp.int32),
                 ],
             )
 
@@ -922,7 +922,7 @@ def _get_weights_func_for_shock(*, name: str, grid: _ShockGrid) -> UserFunction:
             input=transition_probs,
             coordinates=[
                 jnp.full(grid.n_points, fill_value=coordinate),
-                jnp.arange(grid.n_points),
+                jnp.arange(grid.n_points, dtype=jnp.int32),
             ],
         )
